Don't expand merges in ChangeLog
By default, git whatchanged will expand every entry in a merge if a
range of revision numbers is specified. If someone pulls in a large set
of changes from another repository when and that merge or another merge
in the same sync breaks the build then Jenkins could end up sending a
"Build Failed" email to every commiter in the merged repository.

This change stops the merges being expanded so only the committer of the
merge would be e-mailed.