Fix problem with populating values from config when using pipeline
When loading the job config into the job configuration GUI, the
repo-plugin properties are empty. The config file contains the values,
but the plugin is unable to populate config values from the file.

If the user edits anything at job configuration page and saves, the
empty repo-plugin properties are written to file and are cleared!

This patch fixes the problem by populating the values from the config
file to the GUI.