RepoScm: Allow setting extra environment variables
This is useful to tweak repo behavior, especially when used from
pipelines. For example, this can be used to tweak git lfs behavior, to
enable extra Git logs or to support using ssh agent.

It is now possible to use ssh agent or withEnv() block with repo
checkout in Jenkins pipelines, like in the following code:

  sshAgent(['gerrit-ssh-credentials-id']) {
    checkout([$class: 'RepoScm', manifestRepositoryUrl: 'ssh://server/project',
              extraEnvVars: getContext(hudson.EnvVars)])
  }